# doc/api.yaml
openapi: 3.0.0
info: 
  title: Wasa Text
  description: |-
    An API for a messaging app that allows users
    to send, receive, and manage messages and groups.
  version: 0.0.23
components:
  schemas:
...
    Chat:
      type: object
      description: Details of a chat
      properties:
        chatId:
          type: integer
          minimum: 1
          maximum: 100
          description: Unique ID of a chat
        
        name:
          type: string
          minLength: 1
          maxLength: 20
          pattern: "^.*$"
          example: "Mario"
          description: Username or group name
        
        photo:
          $ref: "#/components/schemas/User/properties/photo"
        
        isGroup:
          type: boolean
          description: |
            Indicates if the conversation is a group chat
            or one-on-one chat
          example: false
        
        participants:
          type: array
          minItems: 2  
          maxItems: 500
          description: List of group members
          items:
            $ref: "#/components/schemas/User/properties/userId"
        
        messages:
          $ref: "#/components/schemas/Message"
    Message:
      type: object
      description: Details of a message
      properties: 
        messageId:
          type: integer
          minimum: 1
          maximum: 10000
          example: 3
          description: Unique ID of a message
        
        senderId: 
          $ref: "#/components/schemas/User/properties/userId" 
        
        chatId:
          $ref: "#/components/schemas/Chat/properties/chatId"
        
        content:
          type: string
          minLength: 1
          maxLength: 500
          pattern: "^.*$"
          example: "Hi, i'm marco"
          description: Content of the message
        
        photo:
          $ref: "#/components/schemas/User/properties/photo"
        
        status:
          type: string
          enum: [sent, received, read]
          description: >
            The status of the message:
            - `sent`: The message has been sent but not delivered.
            - `received`: The message has been delivered to the recipient's device.
            - `read`: The message has been read by the recipient.
          example: "read"
        
        timestamp:
          type: string
          format: date-time
          minLength: 0
          maxLength: 22
          example: "2024-11-02T16:26:24Z"
          description: Details about date and time
        
        isForwarded:
          type: boolean
          default: false
          description: If true, the message was forwarded
          example: true

        reactions:
          type: array
          description: List of reactions to a message
          items:
            type: object
            properties:
              userId:
                $ref: "#/components/parameters/UserId"
              reaction:
                type: string
                description: Reaction emoji
                example: "🔥"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: plane
  parameters:
    ChatId:
      schema:
        type: integer
        minimum: 1
        maximum: 1000
      description: Unique chat id
      required: true
      in: path
      name: ChatId
    MsgId:
      schema:
        type: integer
        minimum: 1
        maximum: 10000
      description: Unique msg id
      required: true
      in: path
      name: MsgId
    UserId:
      schema:
        type: integer 
        minimum: 1
        maximum: 1000
        example: 2
       
      description: Unique user id
      required: true
      in: path
      name: UserId
...
paths:
...
  /chats:
    get:
      tags: 
        - Chat
      summary: Allows a user to get a list of his conversations
      description: Allows a user to get a list of his conversations
      operationId: getMyConversations
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of user's conversations
          content:
            application/json:
              schema:
                type: object
                description: Schema of chat's list
                properties:
                  chats:
                    type: array
                    minItems: 0
                    maxItems: 2000
                    description: Array containing all the chats
                    items:
                      $ref: "#/components/schemas/Chat"
      
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                type: object
                description: Schema of error
                properties:
                  error:
                    type: string
                    description: Message of error
                    example: "Invalid request"
    
    post:
      tags:
        - Chat
      summary: Allows a user to create a new Chat
      description: Endpoint to create a new chat
      operationId: newChat
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Chat"
            examples:
              createGroupChat:
                summary: Example of creating a group chat
                value:
                  name: "Project Team"
                  isGroup: True
              createDirectChat:
                summary: Example of creating a direct chat
                value:
                  name: "John Doe"
                  isGroup: False
      
      responses:
        '201':
          description: Chat successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chat'
        
        '400':
          description: Invalid request
        
        '401':
          description: Unauthorized request
  /chats/{ChatId}:
    parameters:
      - $ref: "#/components/parameters/ChatId"
    
    get:
      tags: 
        - Chat
      summary: Allows a user to get the details of a single conversation
      description: Allows a user to get the conversation details
      operationId: getConversation
      security:
        - bearerAuth: []
      responses:
        '200':
          description : details of conversation required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chat"
        
        '400':
          description: Invalid request parameters
        
        '401':
          description: Unauthorized request
          content:
            application/json:
              schema:
                type: object
                description: Schema of error
                properties:
                  error:
                    type: string
                    description: Message of error
                    example: "User not authorized to send a message in this conversation"
        
        '404':
          description: Conversation not found
    
    delete:
      tags:
        - Chat
      summary: Allows a user to delete a chat
      description: deletes a conversation, especially the one specified in the path
      operationId: deleteConversation
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Chat successfully deleted
        
        '401':
          description: Unhautorized request
        
        '404':
          description: Conversation not found
...
  /chats/{ChatId}/messages/{MsgId}/forward:
    parameters:
      - $ref: "#/components/parameters/ChatId" #ID of the conversation containing the original message
      - $ref: "#/components/parameters/MsgId"  #ID of the original message
    
    post:
      tags: 
        - Message
      summary: Forward a message to another chat
      description: Allows a user to forward an existing message to a different chat.
      operationId: forwardMessage
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Schema of chat
              properties:
                targetChatId:
                  $ref: "#/components/schemas/Chat/properties/chatId"
      
      responses:
        '201':
          description: message forwarded successfully
          content:
            application/json:
              schema:
                type: object
                description: Schema of success
                properties:
                  message:
                    $ref: "#/components/schemas/Message"
        
        '400':
          description: Invalid request parameters
        
        '401':
          description: Unauthorized request
          content:
            application/json:
              schema:
                type: object
                description: Schema of error
                properties:
                  error:
                    type: string
                    description: Message of error
                    example: "User not authorized to forward a message in this conversation."
        
        '404':
          description: Conversation or user not found
...
  /chats/{ChatId}/messages/{MsgId}:
    parameters:
      - $ref: "#/components/parameters/ChatId" #ID of the conversation containing the message
      - $ref: "#/components/parameters/MsgId"  #ID of the message to remove
    
    delete:
      tags: 
        - Message
      summary: Allows a user to delete a message
      description: gives the user the possibility to delete a message
      operationId: deleteMessage
      security:
        - bearerAuth: []

      responses:
        '204':
          description: Message deleted successfully
        
        '401':
          description: Unauthorized request
          content:
            application/json:
              schema:
                type: object
                description: Schema of error
                properties:
                  error:
                    type: string
                    description: Message of error
                    example: "User not authorized to remove reaction"
        
        '404':
          description: Message or reaction not found
...
  /chats/{ChatId}/name:
    parameters:
      - $ref: "#/components/parameters/ChatId"
    
    put:
      tags: 
        - Group
      summary: rename a group chat
      description: Allows a user to rename a group chat. The chat must have "isGroup:true"
      operationId: setGroupName
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Shema of group name
              properties:
                name:
                  $ref: "#/components/schemas/Chat/properties/name"
        
      responses:
        '200':
          description: Group name updated successfully
          content:
            application/json:
              schema:
                type: object
                description: Schema of success
                properties:
                  name:
                    type: string
                    description: The new name for the group
        
        '400':
          description: "Invalid input"
        
        '401':
          description: Unauthorized request
          content:
            application/json:
              schema:
                type: object
                description: Schema of error
                properties:
                  error:
                    type: string
                    description: Message of error
                    example: "User not authorized to update group name"
        
        '404':
          description: "Group not found"
  /chats/{ChatId}/photo:
    parameters:
      - $ref: "#/components/parameters/ChatId"
    
    put:
      tags: 
        - Group
      summary: Set or update a new group photo
      description: Allows the user to update the group photo. The chat must have "isGroup:true"
      operationId: setGroupPhoto
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              description: Schema of photo
              properties:
                photo:
                  $ref: "#/components/schemas/User/properties/photo"
      
      responses:
        '200':
          description: Photo updated successfully
          content:
            multipart/form-data:
              schema:
                type: object
                description: Schema of photo
                properties:
                  photo:
                    $ref: "#/components/schemas/User/properties/photo"
        
        '400':
          description: Invalid request
        
        '401':
          description: Unauthorized request
          content:
            application/json:
              schema:
                type: object
                description: Schema of error
                properties:
                  error:
                    type: string
                    description: Message of error
                    example: "User not authorized to update group photo"